+++ src/components/core/skeleton/App/App.stories.jsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 
 import App from './index';
 
-const title = 'Components/App';
+const title = 'Components/Core/Skeleton/App';
 const component = <App />;
 
 storiesOf(title, module).add('default', () => component);
