+++ src/components/core/skeleton/App/__tests__/App.test.jsx
